+++ bin/templates/cordova/lib/prepare.js
@@ -21,7 +21,7 @@ const fs = require('fs-extra');
 const path = require('path');
 const { ConfigParser, xmlHelpers, events, CordovaError } = require('cordova-common');
 const ManifestJsonParser = require('../../../../lib/ManifestJsonParser');
-const PackageJsonParser = require('./PackageJsonParser');
+const PackageJsonParser = require('../../../../lib/PackageJsonParser');
 const SettingJsonParser = require('../../../../lib/SettingJsonParser');
 
 module.exports.prepare = function (cordovaProject, options) {
