+++ lib/PackageJsonParser.js
